refactor according to code climate
