+++ bionc/bionc_numpy/__init__.py
@@ -13,7 +13,8 @@ from .natural_marker import NaturalMarker, Marker, SegmentNaturalVector
 from .natural_segment import NaturalSegment
 from .inertia_parameters import InertiaParameters
 from .enums import JointType
-from .joint import Joint, GroundJoint
+from .joints import Joint
+from .joints_with_ground import GroundJoint
 from .natural_vector import NaturalVector
 from .inverse_kinematics import InverseKinematics
 from .external_force import ExternalForceList, ExternalForce
